requirements.txt
azureml-sdk
flask
requests
luna-publish-utils==0.5.9
sqlalchemy==1.3.13
pyodbc
azure-keyvault-secrets
azure.identity==1.2.0
azure-storage-blob
pyjwt